updated BookCard foder. removed useEffect function from App.jsx to utils.
