+++ src/App.jsx
@@ -1,6 +1,6 @@
-import React, { useEffect, useState } from 'react';
-import { getAllData } from './util/index';
-import BookCardMain from '/src/components/BookCard/BookCardMain.jsx';
+import React from 'react';
+// import { getAllData } from './util/index';
+import BookCardMain from '/src/components/BookCard/BookCard.jsx';
 import booksData from '/src/data/booksData.js';
 
 import BookDetailsMain from '/src/components/BookDetails/BookDetailsMain.jsx';
@@ -8,19 +8,6 @@ import BookDetailsMain from '/src/components/BookDetails/BookDetailsMain.jsx';
 const URL = 'http://localhost:8000/api/v1/';
 
 function App() {
-	const [message, setMessage] = useState('');
-
-	useEffect(() => {
-		(async () => {
-			const myData = await getAllData(URL);
-			setMessage(myData.data);
-		})();
-
-		return () => {
-			console.log('unmounting');
-		};
-	}, []);
-
 	const cards = booksData.map((item) => {
 		return <BookCardMain key={item.id} {...item} />;
 	});
@@ -31,7 +18,6 @@ function App() {
 
 	return (
 		<div className="container">
-			<h1>{message}</h1>
 			<section className="cards-list">{cards}</section>
 			<section className="books-page">{books}</section>
 		</div>
