+++ src/components/BookCard/BookCard.css
@@ -1,3 +1,11 @@
+.card {
+	display: flex;
+	max-width: 350px;
+	border: 1px solid grey;
+	border-radius: 8px;
+	margin: 10px;
+}
+
 .book-card-cover {
 	width: 110px;
 	margin: 10px;
