--- src/components/BookCard/BookCardMain.module.css
@@ -1,7 +1,0 @@
-.bookCard {
-	display: flex;
-	max-width: 350px;
-	border: 1px solid grey;
-	border-radius: 8px;
-	margin: 10px;
-}
