+++ src/util/index.js
@@ -21,5 +21,18 @@ import axios from "axios";
     }
   };
 
+  const [message, setMessage] = useState('');
+
+	useEffect(() => {
+		(async () => {
+			const myData = await getAllData(URL);
+			setMessage(myData.data);
+		})();
+
+		return () => {
+			console.log('unmounting');
+		};
+	}, []);
+
 
   export {getData, getAllData};
